feature: add crc16 calculation to LetoAPI_V1

+++ include/Utils/crc16.hpp
@@ -6,6 +6,6 @@
 #include <stdint.h>
 #include <stddef.h>
 
-LETO_CORE_EXPORT uint16_t calc_crc16(const void* buf, size_t len);
+LETO_CORE_EXPORT uint16_t calc_crc16(const void* buf, uint32_t len);
 
 #endif
