+++ src/Utils/crc16.cpp
@@ -35,11 +35,11 @@ static const uint16_t crc16_table[256] = {
     0x6E17, 0x7E36, 0x4E55, 0x5E74, 0x2E93, 0x3EB2, 0x0ED1, 0x1EF0
 };
 
-uint16_t calc_crc16(const void* buf, size_t len)
+uint16_t calc_crc16(const void* buf, uint32_t len)
 {
     uint16_t crc = 0xFFFF;
     const uint8_t* p = (const uint8_t*)buf;
-    for (size_t i = 0; i < len; i++) {
+    for (uint32_t i = 0; i < len; i++) {
         crc = (crc << 8) ^ crc16_table[(crc >> 8) ^ p[i]];
     }
     return crc;
